pipecat_integration/requirements.txt
python-dotenv==1.1.0
fastapi==0.115.13
uvicorn==0.34.3
pipecat-ai-small-webrtc-prebuilt==0.7.0
#pipecat-ai[sambanova, webrtc, cartesia, silero, daily]